requirements/requirements.txt
graphene==2.1.8
djangorestframework==3.10.0
